+++ Contracts/Responses/ItemPriceResponse.cs
@@ -1,7 +1,7 @@
 namespace BasketManagementAPI.Contracts.Responses;
 
 public sealed record ItemPriceResponse(
-    decimal LineTotal,
-    decimal VatAmount,
-    decimal TotalWithVat);
+    int LineTotal,
+    int VatAmount,
+    int TotalWithVat);
